Add support for single table inheritance classes
