+++ VERSION
@@ -1,1 +1,1 @@
-1.3.1
+1.3.2
